[bg-67] Accurate stop improvement, 'precise-time' mode implemented
Additionally to the "accurate stop" patch, the 'precise-time' stop mode is implemented. It adds additional check of the last buffer's pts against segment stop and ensures precise stop time. Also, type of the 'accurate-stop' property is changed to enum with 3 options: 'disabled', 'any-frame' (same as before) and new mode 'precise-time'.